# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/applicants/{applicant}": {
            "get": {
                "tags": [
                    "Applicants"
                ],
                "summary": "Get a specific applicant",
                "operationId": "91f6661063b99f525c3ab18f8af031c6",
                "parameters": [
                    {
                        "name": "applicant",
                        "in": "path",
                        "description": "Applicant ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ApplicantResource"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Applicant not found"
                    }
                }
            }
        },
...
    },
...
}